edited httpRequests in SpringConfig, renamed payload folder
