+++ src/main/java/nl/novi/bachwtechiteasy/config/SpringSecurityConfig.java
@@ -43,26 +43,20 @@ public class SpringSecurityConfig {
                 .cors(Customizer.withDefaults())
                 .authorizeHttpRequests(auth ->
                         auth
+                                .requestMatchers("**/authorities/**").hasRole("ADMIN")
                                 .requestMatchers(HttpMethod.POST, "/users").hasRole("ADMIN")
                                 .requestMatchers(HttpMethod.GET, "/users").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.POST, "/users/**").hasRole("ADMIN")
+                                .requestMatchers(HttpMethod.GET, "/users/**").authenticated()
+                                .requestMatchers(HttpMethod.PUT, "/users/**").hasRole("ADMIN")
                                 .requestMatchers(HttpMethod.DELETE, "/users/**").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.POST,"/ci-modules").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.PUT,"/ci-modules").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.DELETE,"/ci-modules").hasRole("ADMIN")
-                                .requestMatchers("/ci-modules").authenticated()
-                                .requestMatchers(HttpMethod.POST,"/wall-brackets").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.PUT,"/wall-brackets").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.DELETE,"/wall-brackets").hasRole("ADMIN")
-                                .requestMatchers("/wall-brackets").authenticated()
-                                .requestMatchers(HttpMethod.POST,"/remotes").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.PUT,"/remotes").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.DELETE,"/remotes").hasRole("ADMIN")
-                                .requestMatchers("/remotes").authenticated()
-                                .requestMatchers(HttpMethod.POST,"/televisions").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.PUT,"/televisions").hasRole("ADMIN")
-                                .requestMatchers(HttpMethod.DELETE,"/televisions").hasRole("ADMIN")
-                                .requestMatchers("/televisions").authenticated()
+                                .requestMatchers(HttpMethod.GET,"/ci-modules/**").authenticated()
+                                .requestMatchers("/ci-modules/**").hasRole("ADMIN")
+                                .requestMatchers(HttpMethod.GET,"/wall-brackets/**").authenticated()
+                                .requestMatchers("/wall-brackets/**").hasRole("ADMIN")
+                                .requestMatchers(HttpMethod.GET,"/remotes/**").authenticated()
+                                .requestMatchers("/remotes/**").hasRole("ADMIN")
+                                .requestMatchers(HttpMethod.GET,"/televisions/**").authenticated()
+                                .requestMatchers("/televisions/**").hasRole("ADMIN")
                                 .requestMatchers("/authenticated").authenticated()
                                 .requestMatchers("/authenticate").permitAll()
                                 .anyRequest().denyAll()
